requirements.txt
python-dotenv==1.0.0
twilio==8.2.0
openai==0.27.6
firebase-admin==6.1.0